+++ projects/js-packages/webpack-config/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.11.4 - 2026-07-15
+### Changed
+- Update package dependencies. [#50510]
+
 ## 3.11.3 - 2026-07-13
 ### Changed
 - Internal updates.